Switch ints that can't be negative to uints (#123)
* use uints

* uint don't need a min=0 validation

* Update codec/utils_test.go


* Update input/elasticapm/internal/modeldecoder/v2/metricset_test.go


* Update model/modelpb/util_test.go


* fix renamed methods calls

* Revert "uint don't need a min=0 validation"

This reverts commit dc27da921a5133a571abf86de90d5aedc2b0eed3.

* ignore lint for < 0 in generated code

---------

+++ model/modeljson/internal/metricset.go
@@ -40,7 +40,7 @@ type MetricsetSample struct {
 
 type Histogram struct {
 	Values []float64 `json:"values"`
-	Counts []int64   `json:"counts"`
+	Counts []uint64  `json:"counts"`
 }
 
 func (h Histogram) isZero() bool {
@@ -48,7 +48,7 @@ func (h Histogram) isZero() bool {
 }
 
 type SummaryMetric struct {
-	Count int64   `json:"value_count"`
+	Count uint64  `json:"value_count"`
 	Sum   float64 `json:"sum"`
 }
 
@@ -83,12 +83,12 @@ func (ms *MetricsetSample) MarshalFastJSON(w *fastjson.Writer) error {
 			if i > 0 {
 				w.RawByte(',')
 			}
-			w.Int64(count)
+			w.Int64(int64(count))
 		}
 		w.RawByte(']')
 	case "summary":
 		w.RawString(`,"value_count":`)
-		w.Int64(ms.Summary.Count)
+		w.Int64(int64(ms.Summary.Count))
 		w.RawString(`,"sum":`)
 		w.Float64(ms.Summary.Sum)
 	default:
@@ -100,7 +100,7 @@ func (ms *MetricsetSample) MarshalFastJSON(w *fastjson.Writer) error {
 }
 
 type AggregatedDuration struct {
-	Count int64
+	Count uint64
 	Sum   time.Duration
 }
